이전 DB 수정내용 복구

+++ src/main/java/kr/or/iei/post/model/service/PostService.java
@@ -218,7 +218,13 @@ public class PostService {
 	                for(int i =0; i < spotList.size(); i++) {
 	                	String spotNo = dao.selectSpotNo(conn);
 	                	spotList.get(i).setSpotNo(spotNo);
-	                    result = dao.insertPostSpot(conn, spotList.get(i), postNo);
+                        result = dao.insertPostSpot(conn, spotList.get(i));
+                        if(result < 1) {
+                            JDBCTemplate.rollback(conn);
+                            soptChk = false;
+                            break;
+                        }
+                        result = dao.insertPostSpotManageMent(conn,postNo,spotNo);
 	                    if(result < 1) {
 	                        JDBCTemplate.rollback(conn);
 	                        soptChk = false;
@@ -299,13 +305,25 @@ public class PostService {
 	//게시글 삭제
 	public int deletePost(String postNo, String delRootPath) {
 		Connection conn = JDBCTemplate.getConnection();
+        ArrayList<Spot> spotList = dao.selectPostSpot(conn, postNo);
 		ArrayList<PostComment> commentList = dao.selectCommentList(conn, postNo);
 		ArrayList<PostFile> fileList = dao.selectPostFileList(conn, postNo);
 		
 		int result = dao.deletePost(conn, postNo);
 		System.out.println(result);
 		if(result > 0) {
-			
+            boolean chk = true;
+            for(Spot spot : spotList) {
+            
+                result = dao.deleteSpot(conn, spot.getSpotNo());
+                if(result < 1) {
+                    JDBCTemplate.rollback(conn);
+                    chk = false;
+                    break;
+                }
+            }
+            
+            if(chk) {
 				boolean commentChk = true;
 				for(PostComment comment : commentList) {
 					result = dao.deleteComment(conn, comment.getCommentId());
@@ -328,6 +346,7 @@ public class PostService {
 					}
 					JDBCTemplate.commit(conn);
 				}
+            }
 		}else {
 			JDBCTemplate.rollback(conn);
 		}
@@ -418,6 +437,14 @@ public class PostService {
 				ArrayList<Spot> list = dao.selectPostSpot(conn, post.getPostNo());
 				boolean spotdelChk = true;
 				for (Spot spot : list) {
+                    result = dao.deleteSpotManageMent(conn, spot.getSpotNo());
+                    if (result < 1) {
+
+                        JDBCTemplate.rollback(conn);
+                        spotdelChk = false;
+                        break;
+                    }
+					
 					result = dao.deleteSpot(conn, spot.getSpotNo());
 					if (result < 1) {
 
@@ -433,7 +460,14 @@ public class PostService {
 					for (Spot spot : spotList) {
 						String spotNo = dao.selectSpotNo(conn);
 						spot.setSpotNo(spotNo);
-						result = dao.insertPostSpot(conn, spot, post.getPostNo());
+                        result = dao.insertPostSpot(conn, spot);
+                        if (result < 1) {
+                            JDBCTemplate.rollback(conn);
+                            soptChk = false;
+                            break;
+                        }
+                        
+                        result = dao.insertPostSpotManageMent(conn, post.getPostNo(), spotNo);
 						if (result < 1) {
 							JDBCTemplate.rollback(conn);
 							soptChk = false;
